componente bot organizado
