--- client/src/components/Options/Options.css
@@ -1,14 +1,0 @@
-.options-container {
-    display: flex;
-    flex-wrap: wrap;
-  }
-  
-  .option-button {
-    margin: 5px;
-    border-radius: 25px;
-    padding: 8px;
-    border-color: blue;
-    background: transparent;
-    text-align: center;
-  }
-
