--- client/src/components/Quiz/FlashCard.jsx
@@ -1,28 +1,0 @@
-import React, { useState, useEffect } from "react";
-
-import "./Quiz.css";
-
-const FlashCard = ({ question, answer, incrementIndex }) => {
-  const [showAnswer, setShowAnswer] = useState(false);
-
-  useEffect(() => setShowAnswer(false), [question]);
-
-  return (
-    <>
-      <div
-        className="flashcard-container"
-        onClick={() => setShowAnswer(!showAnswer)}
-      >
-        {!showAnswer && question}
-        {showAnswer && answer}
-      </div>
-      {showAnswer && (
-        <button onClick={incrementIndex} className="flashcard-button">
-          Finalizar Pregunta
-        </button>
-      )}
-    </>
-  );
-};
-
-export default FlashCard;
